% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{bts_datasets}
\alias{bts_datasets}
\title{List BTS datasets}
\usage{
bts_datasets(limit = 50, query = NULL)
}
\arguments{
\item{limit}{Number of datasets (default 50)}

\item{query}{Optional name filter}
}
\value{
tibble: id, name, category, description
}
\description{
List BTS datasets
}
