% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bts.R
\name{bts_query}
\alias{bts_query}
\title{Query a BTS dataset}
\usage{
bts_query(
  dataset_id,
  where = NULL,
  select = NULL,
  group = NULL,
  order = NULL,
  limit = 1000,
  offset = 0
)
}
\arguments{
\item{dataset_id}{Socrata dataset ID}

\item{where}{SoQL where clause}

\item{select}{SoQL select clause}

\item{group}{SoQL group by}

\item{order}{SoQL order}

\item{limit}{Max rows (default 1000)}

\item{offset}{Pagination offset}
}
\value{
tibble
}
\description{
Query a BTS dataset
}
